small edit to review migration to comply with postgres data types

+++ backend/db/migrations/20220923223146-create-review.js
@@ -19,7 +19,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.INTEGER(1)
+        type: Sequelize.INTEGER
       },
       createdAt: {
         allowNull: false,
